Added Hangman & added range to interval
Added a functioning hangman game where it chooses a word from an array. Only problem now is that every "play again", the function rolls twice.. Doesn't stop the gameplay but might come in problematic later on. For the interval function, added a range of time between the calls, but sometimes it doesn.t follow it. Strange.. Very strange